functions/requirements.txt
certifi==2018.11.29
chardet==3.0.4
click==6.7
Flask==1.0.2
future==0.17.1
idna==2.7
itsdangerous==0.24
Jinja2==2.10
line-bot-sdk==1.8.0
MarkupSafe==1.0
python-dotenv==0.10.1
requests==2.19.1
urllib3==1.23
Werkzeug==0.14.1
